Regal Silk
The first of these gifts was a huge bronze bell, on the outside of which was told in pictures the story of the centipede. The second was a sword which would always give its owner the victory. The third was a suit of armor so strong that no swords or arrows could go through it.

The last two were the most wonderful of all. One was a roll of silk of any color he wished, and the more he used of the silk the more the roll grew. The other was a bag of rice which never grew less, although he used all he wished for his friends and himself.