Commit done with no internet access
1. Removed redundant items, the following were removed:

Absinthe Green - This was more of an exercise into event shenanigans at the time. A green dye alternative is not really needed, only white, brown, black, and blue are. Black and Brown dye alternatives will remain in.

Albedo White/Alchemical White - This will return, but under a different name, and use a different, cheaper recipe. Someone just has to pitch me an idea.

Gemstone Amalgam - This was a mistake, and has no use

Textures for all of these will be retained, however.

2. Changed pulverize to use cobble instead of jasper, to reduce cost

3. Removed unused generation settings, since the chalk block no longer exists

4. Experimental chalk recipes. Honey now has a niche use in crafting.

5. Iron gall ink and catechu brown now should be supported in potions

6. Made a new ritual that allows for the creation of sacrosanct ground

7. Added a ritual to summon Vexes, allowing for a source of ectoplasm thats a bit more controllable

8. Localizations everywhere