Beginning of uranid animations
Have you not heard of that madman who lit a lantern in the bright morning hours, ran to the market-place, and cried incessantly: "I am looking for God! I am looking for God!"
As many of those who did not believe in God were standing together there, he excited considerable laughter. Have you lost him, then? said one. Did he lose his way like a child? said another. Or is he hiding? Is he afraid of us? Has he gone on a voyage? or emigrated? Thus they shouted and laughed. The madman sprang into their midst and pierced them with his glances.

"Where has God gone?" he cried. "I shall tell you. We have killed him - you and I. We are his murderers. But how have we done this? How were we able to drink up the sea? Who gave us the sponge to wipe away the entire horizon? What did we do when we unchained the earth from its sun? Whither is it moving now? Whither are we moving now? Away from all suns? Are we not perpetually falling? Backward, sideward, forward, in all directions? Is there any up or down left? Are we not straying as through an infinite nothing? Do we not feel the breath of empty space? Has it not become colder? Is it not more and more night coming on all the time? Must not lanterns be lit in the morning? Do we not hear anything yet of the noise of the gravediggers who are burying God? Do we not smell anything yet of God's decomposition? Gods too decompose. God is dead. God remains dead. And we have killed him. How shall we, murderers of all murderers, console ourselves? That which was the holiest and mightiest of all that the world has yet possessed has bled to death under our knives. Who will wipe this blood off us? With what water could we purify ourselves? What festivals of atonement, what sacred games shall we need to invent? Is not the greatness of this deed too great for us? Must we not ourselves become gods simply to be worthy of it? There has never been a greater deed; and whosoever shall be born after us - for the sake of this deed he shall be part of a higher history than all history hitherto."

Here the madman fell silent and again regarded his listeners; and they too were silent and stared at him in astonishment. At last he threw his lantern to the ground, and it broke and went out. "I have come too early," he said then; "my time has not come yet. The tremendous event is still on its way, still travelling - it has not yet reached the ears of men. Lightning and thunder require time, the light of the stars requires time, deeds require time even after they are done, before they can be seen and heard. This deed is still more distant from them than the distant stars - and yet they have done it themselves."

It has been further related that on that same day the madman entered divers churches and there sang a requiem. Led out and quietened, he is said to have retorted each time: "what are these churches now if they are not the tombs and sepulchres of God?"