Do not register Moonbell biomes in parallel
BlockMoonbell#addValidMoonbellBiome(Biome) would sometimes
cause a crash on startup because multiple threads were
modifying the list at the same time

The fix here is simple and merely changes the
parallelStream() call for a regular stream() call. There is
no advantage to using a parallel stream here as the time to add
an item to a list is extremely tiny. In fact, the overhead of
queueing biome registrations in parallel far outweighs
the amount of time that can be saved.