Blood of Angels
We are the twilight, and we are the dawn.
We are the silence when all else is gone.
We are the ancient ones, we are the flames.
Once you would pray to us, call us by name.

Ours is the power, and ours is the might.
Ours are the shadows that swallow the light.
Worship us mortals, bow down to our reign!
The blood of the angels, it beats in our veins.

Azazel, Sariel, Koshbiel, Shamsiel,
Ananel, Sethiel, Raziel, Zaqiel!

Azazel, Sariel, Koshbiel, Shamsiel,
Ananel, Sethiel, Raziel, Zaqiel!

Gibborim, Annakim, heroes of old.
In words of the prophets our fate is foretold.
Once we were scattered, in hiding, alone.
Now we have marshaled to reclaim our throne.

We are the twilight, and we are the dawn.
We are the silence when all else is gone.
We are the ancient ones, we are the flames.
Once you would pray to us, call us by name.

Azazel, Sariel, Koshbiel, Shamsiel,
Ananel, Sethiel, Raziel, Zaqiel!

Azazel, Sariel, Koshbiel, Shamsiel,
Ananel, Sethiel, Raziel, Zaqiel!

Azazel, Sariel, Koshbiel, Shamsiel,
Ananel, Sethiel, Raziel, Zaqiel!

Azazel, Sariel, Koshbiel, Shamsiel,
Ananel, Sethiel, Raziel, Zaqiel!